phishing_function/requirements.txt
requests==2.32.3
azure-functions
azure-storage-blob
azure-ai-textanalytics
azure-core
azure-identity
python-dotenv
werkzeug 
